+++ tests/integration/RequestBuilder/CampaignRequestBuilderTest.php
@@ -4,8 +4,11 @@ namespace Lmc\Matej\IntegrationTests\RequestBuilder;
 
 use Lmc\Matej\Exception\LogicException;
 use Lmc\Matej\IntegrationTests\IntegrationTestCase;
+use Lmc\Matej\Model\Command\ItemItemRecommendation;
+use Lmc\Matej\Model\Command\ItemUserRecommendation;
 use Lmc\Matej\Model\Command\Sorting;
-use Lmc\Matej\Model\Command\UserRecommendation;
+use Lmc\Matej\Model\Command\UserItemRecommendation;
+use Lmc\Matej\Model\Command\UserUserRecommendation;
 
 /**
  * @covers \Lmc\Matej\RequestBuilder\CampaignRequestBuilder
@@ -30,10 +33,11 @@ class CampaignRequestBuilderTest extends IntegrationTestCase
         $response = static::createMatejInstance()
             ->request()
             ->campaign()
-            ->addRecommendation($this->createRecommendationCommand('a'))
+            ->addRecommendation($this->createUserItemRecommendationCommand('a'))
+            ->addRecommendation($this->createUserUserRecommendationCommand('b'))
             ->addRecommendations([
-                $this->createRecommendationCommand('b'),
-                $this->createRecommendationCommand('c'),
+                $this->createItemItemRecommendationCommand('c'),
+                $this->createItemUserRecommendationCommand('d'),
             ])
             ->addSorting($this->createSortingCommand('a'))
             ->addSortings([
@@ -42,15 +46,36 @@ class CampaignRequestBuilderTest extends IntegrationTestCase
             ])
             ->send();
 
-        $this->assertResponseCommandStatuses($response, ...$this->generateOkStatuses(6));
+        $this->assertResponseCommandStatuses($response, ...$this->generateOkStatuses(7));
     }
 
-    private function createRecommendationCommand(string $letter): UserRecommendation
+    private function createUserItemRecommendationCommand(string $letter): UserItemRecommendation
     {
-        return UserRecommendation::create('user-' . $letter, 'integration-test-scenario')
+        return UserItemRecommendation::create('user-' . $letter, 'integration-test-scenario')
             ->setCount(1)
             ->setRotationRate(1)
             ->setRotationTime(3600);
+    }
+
+    private function createUserUserRecommendationCommand(string $letter): UserUserRecommendation
+    {
+        return UserUserRecommendation::create('user-' . $letter, 'integration-test-scenario')
+            ->setCount(1)
+            ->setRotationRate(1)
+            ->setRotationTime(3600);
+    }
+
+    private function createItemUserRecommendationCommand(string $letter): ItemUserRecommendation
+    {
+        return ItemUserRecommendation::create('item-' . $letter, 'integration-test-scenario')
+            ->setCount(1)
+            ->setAllowSeen(true);
+    }
+
+    private function createItemItemRecommendationCommand(string $letter): ItemItemRecommendation
+    {
+        return ItemItemRecommendation::create('item-' . $letter, 'integration-test-scenario')
+            ->setCount(1);
     }
 
     private function createSortingCommand(string $letter): Sorting