+++ tests/unit/Model/Command/UserItemRecommendationTest.php
@@ -2,19 +2,19 @@
 
 namespace Lmc\Matej\Model\Command;
 
-use Lmc\Matej\Model\Command\Constants\MinimalRelevance;
+use Lmc\Matej\Model\Command\Constants\ItemMinimalRelevance;
 use PHPUnit\Framework\TestCase;
 
-class UserRecommendationTest extends TestCase
+class UserItemRecommendationTest extends TestCase
 {
     /** @test */
     public function shouldBeInstantiableViaNamedConstructorWithDefaultValues(): void
     {
-        $command = UserRecommendation::create('user-id', 'test-scenario');
+        $command = UserItemRecommendation::create('user-id', 'test-scenario');
 
         $this->assertEquals(
             [
-                'type' => 'user-based-recommendations',
+                'type' => 'user-item-recommendations',
                 'parameters' => [
                     'user_id' => 'user-id',
                     'scenario' => 'test-scenario',
@@ -35,12 +35,11 @@ class UserRecommendationTest extends TestCase
         $rotationTime = random_int(1, 86400);
         $modelName = 'test-model-' . md5(microtime());
 
-        $command = UserRecommendation::create($userId, $scenario)
+        $command = UserItemRecommendation::create($userId, $scenario)
             ->setCount($count)
             ->setRotationRate($rotationRate)
-            ->setRotationTime($rotationTime);
-
-        $command->setMinimalRelevance(MinimalRelevance::HIGH())
+            ->setRotationTime($rotationTime)
+            ->setMinimalRelevance(ItemMinimalRelevance::HIGH())
             ->enableHardRotation()
             ->setFilters(['foo = bar', 'baz = ban'])
             ->setModelName($modelName)
@@ -51,7 +50,7 @@ class UserRecommendationTest extends TestCase
 
         $this->assertEquals(
             [
-                'type' => 'user-based-recommendations',
+                'type' => 'user-item-recommendations',
                 'parameters' => [
                     'user_id' => $userId,
                     'count' => $count,
@@ -59,7 +58,7 @@ class UserRecommendationTest extends TestCase
                     'rotation_rate' => $rotationRate,
                     'rotation_time' => $rotationTime,
                     'hard_rotation' => true,
-                    'min_relevance' => MinimalRelevance::HIGH,
+                    'min_relevance' => ItemMinimalRelevance::HIGH,
                     'filter' => 'foo = bar and baz = ban',
                     'properties' => ['item_url'],
                     'model_name' => $modelName,
@@ -77,7 +76,7 @@ class UserRecommendationTest extends TestCase
     /** @test */
     public function shouldAssembleMqlFilters(): void
     {
-        $command = UserRecommendation::create('user-id', 'test-scenario');
+        $command = UserItemRecommendation::create('user-id', 'test-scenario');
 
         // Default filter
         $this->assertArrayNotHasKey('filter', $command->jsonSerialize()['parameters']);
@@ -109,7 +108,7 @@ class UserRecommendationTest extends TestCase
     /** @test */
     public function shouldAllowModificationOfResponseProperties(): void
     {
-        $command = UserRecommendation::create('user-id', 'test-scenario');
+        $command = UserItemRecommendation::create('user-id', 'test-scenario');
         $command->addResponseProperty('test');
         $this->assertSame(['test'], $command->jsonSerialize()['parameters']['properties']);
 
@@ -125,7 +124,7 @@ class UserRecommendationTest extends TestCase
     /** @test */
     public function shouldResetBoostRules(): void
     {
-        $command = UserRecommendation::create('user-id', 'test-scenario')
+        $command = UserItemRecommendation::create('user-id', 'test-scenario')
             ->addBoost(Boost::create('valid_to >= NOW()', 1.0));
 
         $command->setBoosts([
@@ -145,7 +144,7 @@ class UserRecommendationTest extends TestCase
     /** @test */
     public function shouldNotIncludeEmptyBoosts(): void
     {
-        $command = UserRecommendation::create('user-id', 'test-scenario')
+        $command = UserItemRecommendation::create('user-id', 'test-scenario')
             ->setBoosts([]);
 
         $this->assertArrayNotHasKey('boost_rules', $command->jsonSerialize()['parameters']);