+++ tests/unit/RequestBuilder/CampaignRequestBuilderTest.php
@@ -6,8 +6,11 @@ use Fig\Http\Message\RequestMethodInterface;
 use Lmc\Matej\Exception\DomainException;
 use Lmc\Matej\Exception\LogicException;
 use Lmc\Matej\Http\RequestManager;
+use Lmc\Matej\Model\Command\ItemItemRecommendation;
+use Lmc\Matej\Model\Command\ItemUserRecommendation;
 use Lmc\Matej\Model\Command\Sorting;
-use Lmc\Matej\Model\Command\UserRecommendation;
+use Lmc\Matej\Model\Command\UserItemRecommendation;
+use Lmc\Matej\Model\Command\UserUserRecommendation;
 use Lmc\Matej\Model\Request;
 use Lmc\Matej\Model\Response;
 use PHPUnit\Framework\TestCase;
@@ -23,23 +26,24 @@ class CampaignRequestBuilderTest extends TestCase
     {
         $builder = new CampaignRequestBuilder();
 
-        $recommendationCommand1 = UserRecommendation::create('userId1', 'scenario1')
+        $recommendationCommand1 = UserItemRecommendation::create('userId1', 'scenario1')
             ->setCount(1)
             ->setRotationRate(1.0)
             ->setRotationTime(600);
 
-        $recommendationCommand2 = UserRecommendation::create('userId2', 'scenario2')
+        $recommendationCommand2 = UserUserRecommendation::create('userId2', 'scenario2')
             ->setCount(2)
-            ->setRotationRate(0.5)
-            ->setRotationTime(700);
+            ->setRotationRate(1.0)
+            ->setRotationTime(600);
 
-        $recommendationCommand3 = UserRecommendation::create('userId3', 'scenario3')
-            ->setCount(3)
-            ->setRotationRate(0.0)
-            ->setRotationTime(800);
+        $recommendationCommand3 = ItemUserRecommendation::create('itemId1', 'scenario3')
+            ->setCount(3);
+
+        $recommendationCommand4 = ItemItemRecommendation::create('itemId2', 'scenario3')
+            ->setCount(4);
 
         $builder->addRecommendation($recommendationCommand1);
-        $builder->addRecommendations([$recommendationCommand2, $recommendationCommand3]);
+        $builder->addRecommendations([$recommendationCommand2, $recommendationCommand3, $recommendationCommand4]);
 
         $sortingCommand1 = Sorting::create('userId1', ['itemId1', 'itemId2']);
         $sortingCommand2 = Sorting::create('userId2', ['itemId2', 'itemId3']);
@@ -56,10 +60,11 @@ class CampaignRequestBuilderTest extends TestCase
         $this->assertSame('/campaign', $request->getPath());
 
         $requestData = $request->getData();
-        $this->assertCount(6, $requestData);
+        $this->assertCount(7, $requestData);
         $this->assertContains($recommendationCommand1, $requestData);
         $this->assertContains($recommendationCommand2, $requestData);
         $this->assertContains($recommendationCommand3, $requestData);
+        $this->assertContains($recommendationCommand4, $requestData);
         $this->assertContains($sortingCommand1, $requestData);
         $this->assertContains($sortingCommand2, $requestData);
         $this->assertContains($sortingCommand3, $requestData);
@@ -84,7 +89,7 @@ class CampaignRequestBuilderTest extends TestCase
 
         for ($i = 0; $i < 501; $i++) {
             $builder->addRecommendation(
-                UserRecommendation::create('userId1', 'scenario1')
+                UserItemRecommendation::create('userId1', 'scenario1')
                     ->setCount(1)
                     ->setRotationRate(1.0)
                     ->setRotationTime(600)
@@ -122,7 +127,7 @@ class CampaignRequestBuilderTest extends TestCase
         $builder->setRequestManager($requestManagerMock);
 
         $builder->addRecommendation(
-            UserRecommendation::create('userId1', 'scenario1')
+            UserItemRecommendation::create('userId1', 'scenario1')
                 ->setCount(1)
                 ->setRotationRate(1.0)
                 ->setRotationTime(600)